Solve exercise 4
I created a Parser module responsible for transforming each line of the
input into a tuple containing the command name and the list of
arguments.

Each of these tuples is handled by the Runner module, simply by calling
a function in the Runner class with the same name of the command
prefixed by an underscore.